+++ source/conf.py
@@ -34,12 +34,11 @@ extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.coverage',
     'sphinx.ext.napoleon',
-#    'm2r2',
+    'sphinxcontrib.bibtex',
 ]
 
 source_suffix = ['.rst']
-# Enable with m2r2 extention
-#source_suffix = ['.rst', '.md']
+bibtex_bibfiles = ['references.bib']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -61,3 +60,5 @@ html_theme = 'sphinx_rtd_theme'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+
+# -- Options for LATEX output -------------------------------------------------